GPUCP-280 | Favorite wiki profile field
Set a limit on the wikivariables query, so that it returns at least all
the current hydra wikis (~2000).  Cache that result for 4 hours to help
with any perf issues that query causes.

Instead of a type => hidden form for the favwiki preference, use CSS to
hide it and have it be a functional form field, so it updates properly
when the value is changed.

Adjust the preferences js to use the correct elements for setting
values.

Don't load the preferences js module on mobile so we don't break the
rest of preferences js; jquery.ui.autocomplete doesn't exist in
mobilefrontend.

Adjust favorite wiki search api to search the whole domain to support
the use case of typing e.g. "wow.gamepedia.com" and expecting a result.