Add proper robots policy to the UserProfile pages
Currently, we are correctly adding those meta tags to the UserProfiles for the users that are found. Unfortunately, we render different page if the user doesn't exists and which pages are currently indexed. This is not ideal from SEO perspective as those pages don't bring any value and also uses a lot of crawling capacity for the Google, Bing indexers.